+++ Frends.SFTP.ListFiles/Frends.SFTP.ListFiles.Tests/ConnectivityTests.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.IO;
 using System.Threading;
+using System.Threading.Tasks;
 using NUnit.Framework;
 using Frends.SFTP.ListFiles.Definitions;
 using Frends.SFTP.ListFiles.Enums;
@@ -11,11 +12,11 @@ namespace Frends.SFTP.ListFiles.Tests;
 public class ConnectivityTests : ListFilesTestBase
 {
     [Test]
-    public void ListFile_TestWithPrivateKeyFileRsa()
+    public async Task ListFile_TestWithPrivateKeyFileRsa()
     {
         var connection = Helpers.GetSftpConnection();
         connection.Authentication = AuthenticationType.UsernamePasswordPrivateKeyFile;
-        connection.PrivateKeyFilePassphrase = "passphrase";
+        connection.PrivateKeyPassphrase = "passphrase";
         connection.PrivateKeyFile = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "../../../Volumes/ssh_host_rsa_key");
 
         var input = new Input
@@ -26,18 +27,18 @@ public class ConnectivityTests : ListFilesTestBase
             IncludeSubdirectories = false,
             FileEncoding = FileEncoding.ANSI
         };
-        var result = SFTP.ListFiles(input, connection, new CancellationToken());
+        var result = await SFTP.ListFiles(input, connection, new CancellationToken());
         Assert.AreEqual(3, result.FileCount);
     }
 
     [Test]
-    public void ListFile_TestWithPrivateKeyFileRsaFromString()
+    public async Task ListFile_TestWithPrivateKeyFileRsaFromString()
     {
         var key = File.ReadAllText(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "../../../Volumes/ssh_host_rsa_key"));
 
         var connection = Helpers.GetSftpConnection();
         connection.Authentication = AuthenticationType.UsernamePasswordPrivateKeyString;
-        connection.PrivateKeyFilePassphrase = "passphrase";
+        connection.PrivateKeyPassphrase = "passphrase";
         connection.PrivateKeyString = key;
 
         var input = new Input
@@ -48,17 +49,17 @@ public class ConnectivityTests : ListFilesTestBase
             IncludeSubdirectories = false,
             FileEncoding = FileEncoding.ANSI
         };
-        var result = SFTP.ListFiles(input, connection, new CancellationToken());
+        var result = await SFTP.ListFiles(input, connection, new CancellationToken());
         Assert.AreEqual(3, result.FileCount);
     }
 
     [Test]
-    public void ListFiles_TestWithInteractiveKeyboardAuthentication()
+    public async Task ListFiles_TestWithInteractiveKeyboardAuthentication()
     {
         var connection = Helpers.GetSftpConnection();
         connection.UseKeyboardInteractiveAuthentication = true;
 
-        var result = SFTP.ListFiles(_input, connection, new CancellationToken());
+        var result = await SFTP.ListFiles(_input, connection, new CancellationToken());
         Assert.AreEqual(3, result.FileCount);
     }
 }