+++ Frends.SFTP.ListFiles/Frends.SFTP.ListFiles.Tests/ErrorTests.cs
@@ -25,7 +25,7 @@ public class ErrorTests : ListFilesTestBase
         };
             
 
-        var ex = Assert.Throws<SftpPathNotFoundException>(() => SFTP.ListFiles(input, _connection, new CancellationToken()));
+        var ex = Assert.ThrowsAsync<SftpPathNotFoundException>(async () => await SFTP.ListFiles(input, _connection, new CancellationToken()));
         Assert.AreEqual($"No such file", ex.Message);
 
     }
@@ -44,7 +44,7 @@ public class ErrorTests : ListFilesTestBase
             FileEncoding = FileEncoding.ANSI
         };
 
-        Assert.Throws<SocketException>(() => SFTP.ListFiles(input, connection, new CancellationToken()));
+        Assert.ThrowsAsync<SocketException>(async () => await SFTP.ListFiles(input, connection, new CancellationToken()));
     }
 
     [Test]
@@ -63,7 +63,7 @@ public class ErrorTests : ListFilesTestBase
             FileEncoding = FileEncoding.ANSI
         };
 
-        var ex = Assert.Throws<SshAuthenticationException>(() => SFTP.ListFiles(input, connection, new CancellationToken()));
+        var ex = Assert.ThrowsAsync<SshAuthenticationException>(async () => await SFTP.ListFiles(input, connection, new CancellationToken()));
         Assert.AreEqual("Permission denied (password).", ex.Message);
     }
 
@@ -72,7 +72,7 @@ public class ErrorTests : ListFilesTestBase
     {
         var connection = Helpers.GetSftpConnection();
         connection.Authentication = AuthenticationType.UsernamePasswordPrivateKeyFile;
-        connection.PrivateKeyFilePassphrase = "demo";
+        connection.PrivateKeyPassphrase = "demo";
         connection.PrivateKeyFile = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "../../../Volumes/ssh_host_rsa_key");
 
         var input = new Input
@@ -84,7 +84,7 @@ public class ErrorTests : ListFilesTestBase
             FileEncoding = FileEncoding.ANSI
         };
 
-        var ex = Assert.Throws<ArgumentException>(() => SFTP.ListFiles(input, connection, new CancellationToken()));
+        var ex = Assert.ThrowsAsync<ArgumentException>(async () => await SFTP.ListFiles(input, connection, new CancellationToken()));
         Assert.That(ex.Message.StartsWith("Error when initializing connection info:"));
     }
 
@@ -93,7 +93,7 @@ public class ErrorTests : ListFilesTestBase
     {
         var connection = Helpers.GetSftpConnection();
         connection.Authentication = AuthenticationType.UsernamePasswordPrivateKeyFile;
-        connection.PrivateKeyFilePassphrase = "passphrase";
+        connection.PrivateKeyPassphrase = "passphrase";
         connection.PrivateKeyFile = "";
 
         var input = new Input
@@ -105,7 +105,7 @@ public class ErrorTests : ListFilesTestBase
             FileEncoding = FileEncoding.ANSI
         };
 
-        var ex = Assert.Throws<ArgumentException>(() => SFTP.ListFiles(input, connection, new CancellationToken()));
+        var ex = Assert.ThrowsAsync<ArgumentException>(async () => await SFTP.ListFiles(input, connection, new CancellationToken()));
         Assert.That(ex.Message.StartsWith("Error when initializing connection info: "));
     }
 
@@ -116,7 +116,7 @@ public class ErrorTests : ListFilesTestBase
 
         var connection = Helpers.GetSftpConnection();
         connection.Authentication = AuthenticationType.UsernamePasswordPrivateKeyString;
-        connection.PrivateKeyFilePassphrase = "passphrase";
+        connection.PrivateKeyPassphrase = "passphrase";
         connection.PrivateKeyString = key.ToString();
 
         var input = new Input
@@ -128,7 +128,7 @@ public class ErrorTests : ListFilesTestBase
             FileEncoding = FileEncoding.ANSI
         };
 
-        var ex = Assert.Throws<ArgumentException>(() => SFTP.ListFiles(input, connection, new CancellationToken()));
+        var ex = Assert.ThrowsAsync<ArgumentException>(async () => await SFTP.ListFiles(input, connection, new CancellationToken()));
         Assert.That(ex.Message.StartsWith("Error when initializing connection info: "));
     }
 
@@ -149,7 +149,7 @@ public class ErrorTests : ListFilesTestBase
             FileEncoding = FileEncoding.ANSI
         };
 
-        var ex = Assert.Throws<SshConnectionException>(() => SFTP.ListFiles(input, connection, new CancellationToken()));
+        var ex = Assert.ThrowsAsync<SshConnectionException>(async () => await SFTP.ListFiles(input, connection, new CancellationToken()));
         Assert.AreEqual("Key exchange negotiation failed.", ex.Message);
     }
 
@@ -169,7 +169,7 @@ public class ErrorTests : ListFilesTestBase
             FileEncoding = FileEncoding.ANSI
         };
 
-        var ex = Assert.Throws<ArgumentException>(() => SFTP.ListFiles(input, connection, new CancellationToken()));
+        var ex = Assert.ThrowsAsync<ArgumentException>(async () => await SFTP.ListFiles(input, connection, new CancellationToken()));
         Assert.That(ex.Message.StartsWith("Error when initializing connection info: "));
     }
 
@@ -189,7 +189,7 @@ public class ErrorTests : ListFilesTestBase
             FileEncoding = FileEncoding.ANSI
         };
 
-        var ex = Assert.Throws<ArgumentException>(() => SFTP.ListFiles(input, connection, new CancellationToken()));
+        var ex = Assert.ThrowsAsync<ArgumentException>(async () => await SFTP.ListFiles(input, connection, new CancellationToken()));
         Assert.That(ex.Message.StartsWith("Error when initializing connection info: "));
     }
 }